refactor CollectdHost class
this extracts common functionality into an abstract base class. This way the
CollectdHost class itself is less complex as it only needs to provide the HTML
builder methods and restricts its functionality to that responsibility. There
is also now a CollectdHostWithApache class which extends CollectdHost and
provides the added functionality of rendering apache collectd graphs.

The resulting classes are now way less complex and there is a way to add a new
class for something other than CollectD by just extending the base Host class.

In addition this also makes phpmd and codesniffer run clean.